requirements.txt
python_gnupg
python_debian
